pyproject.toml
[project]
name = "scanpi"
version = "0.4.0"
description = "Multi-SDR neighborhood scanner — P25 trunking, GMRS/marine/fire, keyword alerts, hybrid search, geo overlay, accessible over Tailscale"
requires-python = ">=3.11"
license = {text = "GPL-3.0-or-later"}
authors = [{name = "Patrick Ryan"}]
readme = "README.md"

# Core deps installed on both lite and full.
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "aiosqlite>=0.20",
    "numpy>=1.26,<2.0",
    "pydantic>=2.0",
    "paho-mqtt>=2.1",
    "pyyaml>=6.0",
    "requests>=2.32",
    "tomli; python_version < '3.11'",
]

[project.optional-dependencies]
# Transcription — required for whisper output, optional in build-time terms
# because the install scripts pull these via pip directly with target-aware
# CUDA selection.
transcribe = ["faster-whisper>=1.0"]
# Semantic search — bge-small-en + onnxruntime + tokenizers.
# Lite users opt-in via SCANPI_FEATURE_SEMANTIC_SEARCH=1 or profile flag.
semantic = ["onnxruntime>=1.18", "tokenizers>=0.19", "huggingface-hub>=0.24"]
# Watchlist YAML round-trip preservation (preserves comments on save).
yaml-roundtrip = ["ruamel.yaml>=0.18"]
# Cross-channel speaker correlation (full opt-in).
speaker = ["speechbrain>=1.0"]
# Dev / test
dev = ["pytest>=8", "httpx>=0.27"]

[project.scripts]
scanpi = "scanpi.cli:main"
scanpi-v3 = "scanpi.cli_v3:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
scanpi = [
    "web/*.html",
    "web/*.css",
    "web/*.svg",
    "tools/*/page.html",
    "tools/geo/data/*.csv",
]
